# openapi.yaml
openapi: 3.1.0
info:
  title: DriveMeMaybeAPI
  version: 0.1.0
  description: |
    REST API for DriveMeMaybe, a mobile app for tracking shared trips and keeping
    driving balances fair. Manages users, trip groups (invite-by-code), group members,
    trips (driver + passengers, distance and/or duration), per-member balances and
    group driving-time insights. Consumed primarily by the DriveMeMaybe mobile app,
    usable by any HTTP client.
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Development (default Spring Boot port)
  - url: https://api.example.com
    description: Production (replace with real host)
...
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Get current user details
      operationId: getUserProfile
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
              examples:
                default:
                  value:
                    id: 11111111-1111-1111-1111-111111111111
                    email: alice@example.com
                    name: Alice
                    nickname: ali
                    pfp: https://example.com/pfp/alice.png
                    preferences:
                      isDarkMode: true
                      preferredLang: English
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/prefs:
    get:
      tags: [Users]
      summary: List available preference filters (languages)
      operationId: getUserPreferencesFilters
      responses:
        '200':
          description: Languages the client can offer for preferredLang, sorted by code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesFilters'
              examples:
                default:
                  value:
                    languageSelection:
                      - code: DE
                        formalName: Deutsch
                      - code: EN
                        formalName: English
                      - code: FR
                        formalName: Français
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/groups:
    get:
      tags: [Users]
      summary: List groups the current user belongs to
      operationId: listUserGroups
      responses:
        '200':
          description: Groups of the current user (empty list if none).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupDTO'
              examples:
                default:
                  value:
                    - id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                      name: Weekend road crew
                      groupCode: XK7Q2M
                      groupMembers: 2
                      groupTrips: 5
                      pfp: https://example.com/pfp/group.png
                    - id: bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb
                      name: Commute buddies
                      groupCode: Q9W4TA
                      groupMembers: 3
                      groupTrips: 0
                      pfp: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/groups/join/{groupCode}:
    get:
      tags: [Users]
      summary: Join a group by invite code
      operationId: joinGroupByCode
      parameters:
        - $ref: '#/components/parameters/GroupCodePath'
      responses:
        '200':
          description: Joined group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
              examples:
                default:
                  value:
                    id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                    name: Weekend road crew
                    groupCode: XK7Q2M
                    groupMembers: 2
                    groupTrips: 5
                    pfp: https://example.com/pfp/group.png
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/update:
    put:
      tags: [Users]
      summary: Update current user profile
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            examples:
              default:
                value:
                  email: alice.new@example.com
                  name: Alice Smith
                  nickname: ali-s
                  pfp: https://example.com/pfp/alice-new.png
                  preferences:
                    isDarkMode: true
                    preferredLang: EN
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
              examples:
                default:
                  value:
                    id: 11111111-1111-1111-1111-111111111111
                    email: alice.new@example.com
                    name: Alice Smith
                    nickname: ali-s
                    pfp: https://example.com/pfp/alice-new.png
                    preferences:
                      isDarkMode: true
                      preferredLang: English
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/delete:
    delete:
      tags: [Users]
      summary: Delete current user
      operationId: deleteUser
      responses:
        '200':
          description: User deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /trips/{groupId}/trip:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    post:
      tags: [Trips]
      summary: Record a trip in a group
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateDTO'
            examples:
              default:
                value:
                  driver: 11111111-1111-1111-1111-111111111111
                  date: '2026-09-17'
                  durationMinutes: 45
                  distanceKm: 50
                  origin: Berlin
                  destination: Potsdam
                  notes: Evening return
                  passengers:
                    - 22222222-2222-2222-2222-222222222222
      responses:
        '200':
          description: Created trip (API returns 200, not 201).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDTO'
              examples:
                default:
                  value:
                    id: cccccccc-cccc-cccc-cccc-cccccccccccc
                    driverId: 11111111-1111-1111-1111-111111111111
                    driver: ali
                    date: '2026-09-17'
                    durationMinutes: 45
                    distanceKm: 50
                    origin: Berlin
                    destination: Potsdam
                    notes: Evening return
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /trips/{groupId}/{tripId}:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
      - $ref: '#/components/parameters/TripIdPath'
    get:
      tags: [Trips]
      summary: Get trip details
      operationId: getTripDetail
      responses:
        '200':
          description: The trip.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDTO'
              examples:
                default:
                  value:
                    id: cccccccc-cccc-cccc-cccc-cccccccccccc
                    driverId: 11111111-1111-1111-1111-111111111111
                    driver: ali
                    date: '2026-09-17'
                    durationMinutes: 45
                    distanceKm: 50
                    origin: Berlin
                    destination: Potsdam
                    notes: Evening return
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Trips]
      summary: Update a trip
      operationId: updateTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateDTO'
            examples:
              default:
                value:
                  driver: 11111111-1111-1111-1111-111111111111
                  date: '2026-09-18'
                  durationMinutes: 50
                  distanceKm: 55
                  origin: Berlin
                  destination: Potsdam
                  notes: Morning return
                  passengers:
                    - 22222222-2222-2222-2222-222222222222
      responses:
        '200':
          description: Updated trip.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDTO'
              examples:
                default:
                  value:
                    id: cccccccc-cccc-cccc-cccc-cccccccccccc
                    driverId: 11111111-1111-1111-1111-111111111111
                    driver: ali
                    date: '2026-09-18'
                    durationMinutes: 50
                    distanceKm: 55
                    origin: Berlin
                    destination: Potsdam
                    notes: Morning return
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Trips]
      summary: Delete a trip
      operationId: deleteTrip
      responses:
        '200':
          description: Trip deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send `Authorization: Bearer <token>`. Tokens are issued by
        POST /auth/signup, POST /auth/login and POST /auth/google/login
        (1 hour lifetime) and can be re-issued via GET /auth/refresh.
  parameters:
    GroupIdPath:
      name: groupId
      in: path
      required: true
      description: Group UUID.
      schema:
        type: string
        format: uuid
      example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
    TripIdPath:
      name: tripId
      in: path
      required: true
      description: Trip UUID.
      schema:
        type: string
        format: uuid
      example: cccccccc-cccc-cccc-cccc-cccccccccccc
    GroupCodePath:
      name: groupCode
      in: path
      required: true
      description: Human-readable group invite code.
      schema:
        type: string
        minLength: 1
        example: XK7Q2M
  schemas:
    UserDTO:
      type: object
      description: Signup / profile-update payload. `password` and `pfp` are optional. `preferences.preferredLang` is resolved case-insensitively against the languages table (defaults to `EN`); `preferences.isDarkMode` maps to the stored dark-mode flag.
      required: [email, name, nickname]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: secret123
        name:
          type: string
          minLength: 1
          example: Alice
        nickname:
          type: string
          minLength: 1
          example: ali
        pfp:
          type: string
          format: uri
          example: https://example.com/pfp/alice.png
        preferences:
          $ref: '#/components/schemas/UserPreferences'
...
    UserResponseDTO:
      type: object
      description: Public user details. `preferences.preferredLang` is the language formal name (null when unset).
      required: [id, email, name, nickname]
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111111
        email:
          type: string
          format: email
          example: alice@example.com
        name:
          type: string
          example: Alice
        nickname:
          type: string
          example: ali
        pfp:
          type: string
          format: uri
          nullable: true
          example: https://example.com/pfp/alice.png
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      description: User display preferences. `isDarkMode` maps to `PREF_DARK_MODE`; `preferredLang` carries the language code on write and the formal name on read.
      properties:
        isDarkMode:
          type: boolean
          nullable: true
          example: true
        preferredLang:
          type: string
          nullable: true
          example: EN
    UserPreferencesFilters:
      type: object
      description: Available preference options returned by GET /users/prefs. Entries are mapped from the USR_PREFS_LANGUAGES table and sorted by code.
      required: [languageSelection]
      properties:
        languageSelection:
          type: array
          items:
            type: object
            required: [code, formalName]
            properties:
              code:
                type: string
                example: EN
              formalName:
                type: string
                example: English
          example:
            - code: DE
              formalName: Deutsch
            - code: EN
              formalName: English
            - code: FR
              formalName: Français
...
    GroupDTO:
      type: object
      description: Trip group. `pfp` is optional. `groupMembers` / `groupTrips` are derived counts (members.size() / trips.size()).
      required: [id, name, groupCode]
      properties:
        id:
          type: string
          format: uuid
          example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
        name:
          type: string
          minLength: 1
          example: Weekend road crew
        groupCode:
          type: string
          example: XK7Q2M
        groupMembers:
          type: integer
          minimum: 0
          example: 2
        groupTrips:
          type: integer
          minimum: 0
          example: 5
        pfp:
          type: string
          format: uri
          nullable: true
          example: https://example.com/pfp/group.png
...
    TripCreateDTO:
      type: object
      description: Create/update trip payload. Provide distance and/or duration; at least one passenger.
      required: [driver, date, passengers]
      properties:
        driver:
          type: string
          format: uuid
          description: Driver user UUID (must be a group member).
          example: 11111111-1111-1111-1111-111111111111
        date:
          type: string
          format: date
          example: '2026-09-17'
        durationMinutes:
          type: integer
          minimum: 0
          nullable: true
          example: 45
        distanceKm:
          type: integer
          minimum: 0
          nullable: true
          example: 50
        origin:
          type: string
          nullable: true
          example: Berlin
        destination:
          type: string
          nullable: true
          example: Potsdam
        notes:
          type: string
          nullable: true
          example: Evening return
        passengers:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
          example:
            - 22222222-2222-2222-2222-222222222222
    TripDTO:
      type: object
      description: Trip details with resolved driver nickname.
      required: [id, driverId, driver, date]
      properties:
        id:
          type: string
          format: uuid
          example: cccccccc-cccc-cccc-cccc-cccccccccccc
        driverId:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111111
        driver:
          type: string
          description: Driver nickname.
          example: ali
        date:
          type: string
          format: date
          example: '2026-09-17'
        durationMinutes:
          type: integer
          nullable: true
          example: 45
        distanceKm:
          type: integer
          nullable: true
          example: 50
        origin:
          type: string
          nullable: true
          example: Berlin
        destination:
          type: string
          nullable: true
          example: Potsdam
        notes:
          type: string
          nullable: true
          example: Evening return
...
    Error:
      type: object
      description: |
        Central error envelope from CustomExceptionHandler (CustomErrorResponse).
        NOTE: the current handler always constructs the body with NOT_FOUND
        (statusCode 404 / statusName NOT_FOUND) while the HTTP status line
        carries the real code — examples below show the intended contract.
      required: [errorCode, detail, statusCode, statusName, timestamp]
      properties:
        errorCode:
          type: string
          description: Machine-readable code, e.g. validation-exception, authorizationException.
          example: validation-exception
        detail:
          type: string
          example: Group not found
        statusCode:
          type: integer
          example: 404
        statusName:
          type: string
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
          example: '2026-09-17T12:00:00'
  responses:
    BadRequest:
      description: Validation / malformed request (ValidationException → HTTP 400).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: validation-exception
                detail: Group name must not be blank
                statusCode: 400
                statusName: BAD_REQUEST
                timestamp: '2026-09-17T12:00:00'
    Unauthorized:
      description: Authentication failed (AuthenticationException → HTTP 401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: authentication-exception
                detail: Invalid email or password
                statusCode: 401
                statusName: UNAUTHORIZED
                timestamp: '2026-09-17T12:00:00'
    Forbidden:
      description: Missing, invalid or expired JWT (JwtException / BadCredentialsException → HTTP 403 with errorCode authorizationException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: authorizationException
                detail: JWT expired
                statusCode: 403
                statusName: FORBIDDEN
                timestamp: '2026-09-17T12:00:00'
    NotFound:
      description: Entity missing (NoEntitiesFoundException → HTTP 404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: no-entities-found
                detail: Group aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa not found
                statusCode: 404
                statusName: NOT_FOUND
                timestamp: '2026-09-17T12:00:00'
    InternalError:
      description: Server failure (DatabaseException / UnexpectedException / generic fallback → HTTP 500).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: generic-exception
                detail: An unexpected error occurred, we are working on it.
                statusCode: 500
                statusName: INTERNAL_SERVER_ERROR
                timestamp: '2026-09-17T12:00:00'
security:
  - bearerAuth: []